refactor: common/response를 response로 이동
