+++ src/main/java/ex/sample/domain/sample/implementation/SampleReader.java
@@ -3,8 +3,8 @@ package ex.sample.domain.sample.implementation;
 import ex.sample.domain.sample.domain.Sample;
 import ex.sample.domain.sample.domain.SampleRepository;
 import ex.sample.global.annotation.Implementation;
-import ex.sample.global.common.response.ErrorCase;
 import ex.sample.global.exception.GlobalException;
+import ex.sample.global.response.ErrorCase;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
