+++ src/main/java/ex/sample/global/jwt/JwtAuthFilter.java
@@ -1,8 +1,8 @@
 package ex.sample.global.jwt;
 
-import ex.sample.global.common.response.ErrorCase;
 import ex.sample.global.exception.GlobalException;
 import ex.sample.global.redis.RedisUtil;
+import ex.sample.global.response.ErrorCase;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
