+++ src/main/java/ex/sample/global/response/CommonResponse.java
@@ -1,4 +1,4 @@
-package ex.sample.global.common.response;
+package ex.sample.global.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
